LPS-65012 Fix internal package in google-maps
