+++ modules/apps/google-maps/google-maps-web/src/main/java/com/liferay/google/maps/web/internal/upgrade/v1_0_0/UpgradePortletId.java
@@ -12,17 +12,23 @@
  * details.
  */
 
-package com.liferay.google.maps.web.constants;
+package com.liferay.google.maps.web.internal.upgrade.v1_0_0;
+
+import com.liferay.google.maps.web.internal.constants.GoogleMapsPortletKeys;
+import com.liferay.portal.kernel.upgrade.BaseUpgradePortletId;
 
 /**
- * @author Manuel de la Peña
+ * @author Peter Fellwock
  */
-public class GoogleMapsConstants {
+public class UpgradePortletId extends BaseUpgradePortletId {
 
-	public static final String BICYCLING = "BICYCLING";
-
-	public static final String DRIVING = "DRIVING";
-
-	public static final String WALKING = "WALKING";
+	@Override
+	protected String[][] getRenamePortletIdsArray() {
+		return new String[][] {
+			new String[] {
+				"1_WAR_googlemapsportlet", GoogleMapsPortletKeys.GOOGLE_MAPS
+			}
+		};
+	}
 
 }
