+++ modules/apps/google-maps/google-maps-web/src/main/resources/META-INF/resources/init.jsp
@@ -24,7 +24,7 @@ taglib uri="http://liferay.com/tld/theme" prefix="liferay-theme" %><%@
 taglib uri="http://liferay.com/tld/ui" prefix="liferay-ui" %><%@
 taglib uri="http://liferay.com/tld/util" prefix="liferay-util" %>
 
-<%@ page import="com.liferay.google.maps.web.constants.GoogleMapsConstants" %><%@
+<%@ page import="com.liferay.google.maps.web.internal.constants.GoogleMapsConstants" %><%@
 page import="com.liferay.portal.kernel.util.Constants" %><%@
 page import="com.liferay.portal.kernel.util.GetterUtil" %><%@
 page import="com.liferay.portal.kernel.util.PortalUtil" %><%@
